+++ controlplane/proto/controller/gen/pb-go/controller_grpc.pb.go
@@ -24,8 +24,6 @@ const _ = grpc.SupportPackageIsVersion7
 type ControllerClient interface {
 	// Returns the latest configuration of a DoubleZero node based on on-chain data
 	GetConfig(ctx context.Context, in *ConfigRequest, opts ...grpc.CallOption) (*ConfigResponse, error)
-	// Returns only the hash of the latest configuration (lightweight check for changes)
-	GetConfigHash(ctx context.Context, in *ConfigRequest, opts ...grpc.CallOption) (*ConfigHashResponse, error)
 }
 
 type controllerClient struct {
@@ -45,23 +43,12 @@ func (c *controllerClient) GetConfig(ctx context.Context, in *ConfigRequest, opt
 	return out, nil
 }
 
-func (c *controllerClient) GetConfigHash(ctx context.Context, in *ConfigRequest, opts ...grpc.CallOption) (*ConfigHashResponse, error) {
-	out := new(ConfigHashResponse)
-	err := c.cc.Invoke(ctx, "/controller.Controller/GetConfigHash", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
 // ControllerServer is the server API for Controller service.
 // All implementations should embed UnimplementedControllerServer
 // for forward compatibility
 type ControllerServer interface {
 	// Returns the latest configuration of a DoubleZero node based on on-chain data
 	GetConfig(context.Context, *ConfigRequest) (*ConfigResponse, error)
-	// Returns only the hash of the latest configuration (lightweight check for changes)
-	GetConfigHash(context.Context, *ConfigRequest) (*ConfigHashResponse, error)
 }
 
 // UnimplementedControllerServer should be embedded to have forward compatible implementations.
@@ -70,9 +57,6 @@ type UnimplementedControllerServer struct {
 
 func (UnimplementedControllerServer) GetConfig(context.Context, *ConfigRequest) (*ConfigResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetConfig not implemented")
-}
-func (UnimplementedControllerServer) GetConfigHash(context.Context, *ConfigRequest) (*ConfigHashResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetConfigHash not implemented")
 }
 
 // UnsafeControllerServer may be embedded to opt out of forward compatibility for this service.
@@ -104,24 +88,6 @@ func _Controller_GetConfig_Handler(srv interface{}, ctx context.Context, dec fun
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Controller_GetConfigHash_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(ConfigRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ControllerServer).GetConfigHash(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/controller.Controller/GetConfigHash",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ControllerServer).GetConfigHash(ctx, req.(*ConfigRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
 // Controller_ServiceDesc is the grpc.ServiceDesc for Controller service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -132,10 +98,6 @@ var Controller_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "GetConfig",
 			Handler:    _Controller_GetConfig_Handler,
-		},
-		{
-			MethodName: "GetConfigHash",
-			Handler:    _Controller_GetConfigHash_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},